[TASK] do not try to process missing files
It happens, that files are removed from filesystem. Thus sys_file/sys_file_reference remain with a "missing"-flag.
Early exit if file is not available anymore.
Use "is_file" rather than "file_exists". It is the right choice here (and faster too) as symlinks on file-storages are not common at all.